requirements.txt
# Web framework
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
websockets>=13.0
python-multipart>=0.0.12

# Config
pydantic>=2.9.0
pydantic-settings>=2.6.0
python-dotenv>=1.0.0

# LangGraph / LangChain
langgraph>=0.2.50
langchain-openai>=0.2.10
langchain-core>=0.3.0

# AI services
openai>=1.54.0
deepgram-sdk>=3.7.0
elevenlabs>=1.9.0

# Utilities
httpx>=0.27.0
